+++ CHANGES.md
@@ -1,8 +1,20 @@
 ## Change Log
 
-#### 0.1.6
-* Release tasks mustRunAfter instead of dependsOn release preparation task
-* Change project version when finalizing the release preparation task to release version
+#### 0.2.0
+* Changed release tasks to mustRunAfter instead of dependsOn releasePreparation task
+* Change project version when finalizing releasePreparation task to release version
+* Adds 'releaseRefreshArtifacts' task that deals with the dynamic version change of 'releasePrepare'
+* Adds new project properties: 
+** release.version: Release version will be infered from property instead of gradle.properties if set
+** release.interactive: Allows to define whether the 'releasePrepare' task should ask for user feedback
+** release.prepare: Allows to define if 'releasePrepare' should be executed before 'release'
+** release.finalize: Allows to define if 'releaseFinalize' should be executed after 'release'
+* Adds releaseCheck task which checks if release is done from defined master branch
+* Moved tagging of release from 'releasePrepare' to 'release' task
+* Added Gradle 2.1 compatible plugin name 'com.rapidminer.gradle.release'
+
+#### 0.1.5
+* Fixes retrieval of tag message and tag name
 
 #### 0.1.4
 * Fixes wrong declaration of release preperation and release task dependencies